Create Command To Generate CLI Documentation for lokoctl
lokoctl is a growing program, and writing documentation
for code by hand uses time that could be spent
on other tasks. To offset this, we can use cobra to
write documentation automatically.

The command is run as `go run cli/cmd/document/main.go`,
and takes one argument: the path the user wants to output
documentation to.

Requires vendoring cobra/doc/.

Related to kinvolk/PROJECT-Lokomotive-Kubernetes#467