packet: Use user-specified pool name for TF pool
Before this commit, the pool name specified in lokoctl was being
completely ignored. The terraform code just iterated over the worker
pools, and the first one was named "pool-0", second "pool-1", etc.

This can create confusion for the user (an error on "pool-1" when the
name was "foo" is not meaningful, is difficult to know to which pools
servers belong). But, WAY more importantly, if the user reorders the
pools in the lokocfg file (or if for some reason that order is not
guaranteed by golang), it will create a mess in terraform.

This patch just makes that part sane: the terraform worker pools are
created with the user specified name. Therefore, errors in terraform are
meaningful to the user, the user can match servers hostnames in Packet
and can re-order pools without making terraform go nuts.

Also, this patch adds adds a check that all pool names are different.