*: remove flannel support
We don't really use it, it's not supported on Packet and we don't test
it. Let's remove it and add it later if we see demand.

This removes firewall rules for VXLAN for some platforms too. Note that
if we want to support the Calico's VXLAN encapsulation mode
(https://docs.projectcalico.org/getting-started/kubernetes/installation/config-options#switching-from-ip-in-ip-to-vxlan)
in the future we should enable those conditionally.