pkg/terraform: always execute plan with verbose
We want to always show the `terraform plan` diff to the user so they can
make a decision whether to apply the changes or not.

Currently the output of `terraform plan` is either shown or not
depending on the `--verbose` flag, so users get a message like

```
Do you want to proceed with cluster install? [type "yes" to continue]:
```

Without any change context if `--verbose` is false.

This changes Plan() to run `terraform refresh` with user-configured
verbosity and then `terraform plan -refresh=false` showing always the
output.

The effect is that users will see the refreshing logs and the plan diff
if `--verbose` is true and only the plan diff if `--verbose` is false.