prometheus-operator: Add variable `watch_labeled_prometheus_rules`
This commit adds a variable `watch_labeled_prometheus_rules` to the
prometheus component config. Using this config user can decide if they
want to watch all the `PrometheusRule` in the cluster or only those
labeled as `release: "prometheus-operator"`.

This commit sets the value of `ruleSelectorNilUsesHelmValues` in the
values file behind the scene to what user has provided.

When the value of `watch_labeled_prometheus_rules` is `true` the
`Prometheus` object's field `ruleSelector` is set with following:

```yaml
ruleSelector:
  matchLabels:
    release: "prometheus-operator"
```

Due to this any `PrometheusRule` object that does not have above labels,
they won't be selected.

And when `ruleSelectorNilUsesHelmValues` is set to `false`,
`ruleSelector` field in `Prometheus` object is empty.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>