Makefile: Only run linter if code compiles
There are ton of cases where the linter can't run, with errors like:

	WARN [runner] Can't run linter goanalysis_metalinter: assign: failed prerequisites: inspect@github.com/kinvolk/lokomotive/pkg/platform/packet
	WARN [runner] Can't run linter unused: buildssa: analysis skipped: errors in package: [/home/rodrigo/src/kinvolk/lokomotive/pkg/platform/packet/packet.go:271:20: cannot use arguments (variable of type []string) as string value in argument to ex.Execute]
	ERRO Running error: buildssa: analysis skipped: errors in package: [/home/rodrigo/src/kinvolk/lokomotive/pkg/platform/packet/packet.go:271:20: cannot use arguments (variable of type []string) as string value in argument to ex.Execute]

The error is not crystal clear, but if you look closely it says there
were some errors in the package. While not very clear, that seems to
mean **compilation** errors.

I've faced this linter error several times in the last days and kept the
branches to later analyze (didn't find the reason immediately). I can
confirm that all the branches I kept where the linter failed to run with
similar errors, there were compilation errors and running "make lint"
(with this patch) makes it obvious.

Adding this dependency, the error is way clear, as it shows the
compilation error, like:

	pkg/platform/packet/packet.go:270:17: too many arguments in call to ex.Apply

This would have saved me quite some time, as this is way more clear.
And I sometimes trying to fix a linter error end up creating a syntax
error.