e2e: Invoke monitoring tests from single location
This commit adds a test function which initialises the prometheus client
and then passes it to the prometheus tests that need it.

This approach is preferable over creating a new port-forward for every
test is because initialising a port forward takes around 3-4 seconds and
this can significantly increase the total running time of tests. Reusing
port forward connection and prometheus client saves time for test.

And having a single place of creating port forward and client
intialisation ensures that we close that connection gracefully.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>