quickstarts,examples: expose cluster_region and clarify
The AWS quickstart shows how to configure the `AWS_DEFAULT_REGION`
environment variable without explaining that this only applies to the
lokoctl S3 backend.

Also, the current examples don't expose the cluster region so a user
might expect `AWS_DEFAULT_REGION` to change the region where the cluster
is deployed.

This clarifies what `AWS_DEFAULT_REGION` configures in the quickstart
and exposes a new `cluster_region` parameter in the example AWS
configurations.