packet: Add support for reserved instances
This commit adds support for reserved instances, which are already
supported in terraform.

This patch also fixes a long standing bug where the worker pools use
reservations IDs from the controller nodes, instead of workers nodes.
Basically, it was totally broken. This was overseen when creating
template, and the lack of tests for generated templates didn't help.

Fixes: #20