packet: Add support for reserved instances
This commit adds support for reserved instances, as they were ady
supported in terraform.

This patch also fixes a long standing bug where the worker pools 't have
a reservation id but used the controller ones (if they ted).  Basically,
it was totally broken. This was overseen when creating template, and the
lack tests for generated templates didn't help.

Fixes: #20