kube-apiserver: switch all probes to use HTTPS instead of TCP
To avoid TLS handshake errors being logged extensively on Packet.

As we set --anonymous-auth=false and we don't want to change it, we need to
set readiness probes to do authentication. The easiest way of doing that
is via --token-auth-file with static user named nobody. The access token
is generated by Helm by calculating SHA256 of secrets template, meaning
the token will be different for each cluster and will be changed on any
certificate change. The nobody user has no RBAC grants, so it has only
access to /healthz endpoint as far as I know (I checked getting pods and
/metrics too and got 401).

For AWS, the TLS handshake errors cannot be resolved, as they are caused
by AWS Load Balancer, which is configured to only TCP checks and in HTTP
mode, it does not support authentication header, so we would have to
remove --anonymous-auth=false kube-apiserver option, which we shouldn't
do.

Refs #278

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>