terraform: Add ApplyArgs() to reduce duplication
This patch just creates the ApplyArgs() helper to terraform executor, so
code don't need to be duplicating the logic of how to apply
(auto-approve, etc.) and also improves logging.

Without this patch, places calling Execute() wouldn't log, and therefore
when used to apply the user will have to wait (sometimes ~10 minutes)
until something else is shown on the console. By adding this helper, we
can easily log when it is called and the user can be aware that the
operation will take some minutes.