prometheus operator: Add seccomp annotations to PSP
This commit cherry picks the changes made in upstream which adds a
functionality to specify custom annotations to the PSP. Using this
feature one can provide the seccomp annotations on the PSP.
See PR: helm/charts#21897.

Entire chart is not updated because there is a bug in prometheus
operator code that renders the update useless.
See Issue: coreos/prometheus-operator#3093.