kube-apiserver: disable HAProxy checks for backend
As we always have only one backend, it doesn't make sense to perform
those checks. Backend health checks in HAProxy should be used, when
there is multiple backends and you want to exclude unhealthy ones from
receiving traffic. In our case, we only have one backend AND we don't do
any caching or custom error messages via HAProxy, so we can just always
try to pass the traffic to the kube-apiserver, making HAProxy even more
transparent in the setup.

Perhaps in the future, we should add liveness probes for both
containers, so if any of the process gets stuck, kubelet will restart
it, but this is rather out of scope of this commit.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>