packet: Remove optimization to create workers
Creating workers at the same time than the controllers only saves time
if the cluster is small. If you have more than 3 workers, this won't
save much time. Also, it will increase a lot the time the user will wait
to answer the next prompt.

This patch just creates the controllers and asks the user ASAP for the
next input. This way, the user needs to wait for manual steps as little
as possible.

This may increase the total time it takes to create the cluster, but
reduces the time the user needs to different stages and do manual
actions.