refactor: remove install package(install.go)
install package now consists of one file:
- install.go: which deals with configuring terraform.

install package by its name and contents of the files serve no purpose,
hence its files should be moved around to correct packages.

Furthermore the contents of the file i.e functions are utility functions
to prepare terraform modules and directory, something which should be
part of the `pkg/terraform/util.go` and not be a standalone file.

This commit moves the contents of `pkg/install/install.go` to the file
`pkg/terraform/util.go` and updates the references used elsewhere in
the codebase.

Signed-off-by: Imran Pochi <imran@kinvolk.io>