platform/aws: Move LB listener to workers module
NLB listeners are tightly-coupled to the worker pool they serve.
A listener can point at only one worker pool. Since we can have an
arbitrary number of worker pools, we should create per-pool listeners
(one for HTTP and another for HTTPs).
Since we can't create two listeners with the same protocol and port,
we parameterize the ports (with defaults of 80/443 for HTTP/HTTPS,
respectively).