refactor: remove install package(verify.go)
The contents of `verify.go` deals with talking to the Kubernetes API to
verify the health of the Lokomotive cluster.

Since we already have a package `k8sutil` that deals with talking to the
Kubernetes API, package `k8sutil` seems to be a better place for
`verify.go`.

Also this contents of the function seems a better candidate to be a
method of the already existing `Cluster` struct in `k8sutil` package
rather than a standalone function in a separate file.

This commit moves the content of `pkg/install/verify.go` to
`pkg/k8sutil/cluster.go` as a method of `Cluster` struct and updates
the package references used elsewhere in the codebase.

Since `Health` and `Verify` are ambigious, `Health` is now renamed to
`ComponentsStatus` as it deals with returning the status of the
control plane components (for now only etcd).

Signed-off-by: Imran Pochi <imran@kinvolk.io>