refactor: internal package for helper functions.
Adding a new internal package for utility and helper functions that
need not be exposed outside.

Currently there are lot of util packages in the codebase. Having an
internal package helps in consolidating all such helper functions in
this internal package rather than having separate util package/files
in different places in the codebase.

The first function that is to be added is RenderTemplate which renders
provided go template.

Adds unit tests for the same.

Signed-off-by: Imran Pochi <imran@kinvolk.io>