docs/how-to-guides: Add a guide for Cloudflare DNS
Although at the moment Cloudflare is supported only with Packet
clusters, Cloudflare DNS support isn't strictly tied to Packet.
Since the Packet quickstart guide is using Route 53 as the DNS
provider, it makes sense to me to add the Cloudflare-related
information as a how-to guide for the sake of keeping the quickstart
simple and minimal. Putting this information in the configuration
reference seemed strange to me since we aren't really providing
information about config knobs but are rather giving a "recipe" for
achieving a specific result (a cluster with Cloudflare DNS).