test/components/kubernetes: disable kubelet disruptive label test
This commit disables self-hosted kubelet tests for picking up correct
node labels, as it is flaky because of runc bug in current release of
Flatcar stable, which makes kubelet pod to never terminate when the pod
is scheduled for removal.

Once the fix of runc reaches Flatcar stable, this test should be
re-enabled.

See also #110

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>