test/system: initial commit
This commit adds e2e test, to verify, that there is no extra SSH keys
added to the Kubernetes nodes, other than one specified in the SSH_KEY
environment variable, which is used by the CI for provisioning the
clusters.

Ideally, the test should pull the keys directly from the cluster
configuration, however this is currently not trivial to implement.

Refs #465

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>