components: add required labels and annotations for Helm updates
When upgrading a component that had an apiVersion change, Helm would
complain with "rendered manifests contain a new resource that already
exists".

This was fixed in Helm 3.2.0 by adding labels and annotations to mark a
resource as managed by Helm.

This adds those labels and annotations to problematic objects in
Lokomotive components.

We'll probably need to add them to more objects as apiVersions change,
since we'll check upgrades when releasing new versions, we'll handle
those as they happen.

See https://github.com/helm/helm/pull/7625 and
https://github.com/helm/helm/pull/7649.