Packet etcd: Use http for metrics endpoint
Change the metrics port of etcd from `https` to `http` because:

- When you keep metrics port on https you need certificate to scrape
that endpoint. You can't simply skip the TLS check and expect to get the
data, a client cert is needed.

- Providing the apiserver client cert to prometheus operator is
counter productive to security. So it is not a very viable option.
Because this cert has root permissions on the etcd cluster.

- We can create another user that has permissions to scrape metrics
endpoint only, but it is not trivial. See the upstream issue which
mentions how cert access etcd is either access to everything or nothing.
Issue: https://github.com/etcd-io/etcd/issues/8458.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>