test/calico/calico_test.go: drop dependency on libcalico-go
While trying to upgrade Helm dependency to the latest version, we hit
the issue, that libcalico-go is still using old version of
'k8s.io/apimachinery', which prevents us from updating. On Calico slack,
we found, that libcalico-go shouldn't really be imported and used, as
the stability is not guaranteed.

We could upgrade apimachinery in libcalico-go and use the forked version
for the time being, but I think we should avoid that, as implementing
that, getting it merged upstream and released may take a long time.

As TestHostEndpoints test is the only place we use libcalico-go, the
easier way forward is to drop it as a dependency, which will unblock
updating Helm.

This commit changes this test to use Kubernetes REST client, which
mimics 'kubectl get --raw' instead of Calico client, to get HostEndpoint
objects from the cluster.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>