test/calico/calico_test.go: improve test name and logic
This commit changes the TestHostEndpoints test, to validate, that not
any HostEndpoint object is created for all Node objects in the cluster,
but also that it is being created for public interface, which is our
main concern, as if given HostEndpoint object is not created, it may
expose sensitive ports to internet.

This commit also changes the name of this test to reflect this change.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>