cli/cmd: rename --kubeconfig flag to --kubeconfig-file
This also renames environment variable equivalent of old flag, which was
KUBECONFIG to KUBECONFIG_FILE.

This is to avoid prefering KUBECONFIG environment variable over
cluster configuration, which is happening right now, as we treat
--kubeconfig flag and KUBECONFIG environment variable as the same thing,
implicitly via 'viper.AutomaticEnv()' call.

The order of selecting kubeconfig file to use should be following:
1. --kubeconfig-file flag OR KUBECONFIG_FILE environment variable
2. cluster configuration in working directory
3. KUBECONFIG environment variable (currently not implemented)
4. ~/.kube/config file

Part of #595.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>