Don't show release values when running Terraform
This commit adds 'sensitive = true' to all Terraform outputs, so they
don't show up, when we run 'terraform apply' from the code. This is
to avoid cluttering the user output, as those values are information,
which user don't need to see.

As far as I know, adding 'sensitive = true' is the only way in Terraform
to hide this output, even though the values itself are not really
sensitive.

Closes #625

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>