pkg/k8sutil: rename NewClientset to NewClientsetFromFile
This commit renames NewClientset function into NewClientsetFromFile to
better describe it's functionality, as function takes file path as an
argument to build the clientset object.

It also makes room for new NewClientset function, which will take
content of kubeconfig file instead of file path, to make it decoupled
from dependency on local filesystem.

NewClientsetFromFile is only a temporary and should be removed once we
switch back to use NewClientset function everywhere.

This commit also adds some small documentation to NewClientsetFromFile
function to avoid triggering the linter.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>