terraform/aws/workers: rename LB target group
Doing so will cause the LoadBalancer to be recreated by Terraform. First
the new one will be created and then the old one will be destroyed.

We need to do this because our LoadBalancers changed their
health_check.matcher to TCP and modifying them in place is not supported
by the AWS provider.

Also, this uses an underscore instead of a dash to separate words, which
better follows Terraform's conventions[[1]].

[1]: https://www.terraform.io/docs/extend/best-practices/naming.html#resource-names

Co-authored-by: Mateusz Gozdek <mateusz@kinvolk.io>