goreleaser: stop marking releases as pre-releases
Originally the idea with marking releases as pre-releases was that
releases before 1.0.0 should be pre-releases.

However, this doesn't seem to be the intention of the check. It's more
oriented towards alpha or RC releases before a final release is cut.

Marking a release as pre-release also prevents it from appearing on the
main repo page.