refactor: Chart creation
In every component, we "render" the chart to yaml files and then give it
for the installation. Map of file name and file content acts as a common
interface among all the components. This is done because some components
don't have any helm chart at all.

Since we have the common interface as `map[string]string` and we have to
install things using helm we have to create a chart out of these yaml
content again. This is very inefficient for the components that ship
chart because they are going through the process twice.

So if we change the common interface to helm `release` we don't have to
do the double conversion. And the components that don't have any helm
chart are responsible for returning helm release.

This frees the intaller code from creating any helm chart. This also
adds other benefits like the ordering. When we do render-manifest we
lose out on the ordering because our common interface was map, but with
this change we preserve consistent ordering.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>