k8sutil: Utility functions to create/update Namespace
Move function `ensureNamespaceExists` to k8sutil package with new name
CreateOrUpdateNamespace. This function creates/updates the namespace.

Components use this function to create/update the release namespace.

This function is also used to update the `kube-system` namespace after
cluster installation.

Includes another helper function in the `internal` package that merges
two given maps. This is done to ensure that existing labels and
annotations are retained and the new set of Labels/annotations overrides
any existing ones.

Signed-off-by: Imran Pochi <imran@kinvolk.io>