pkg/components/util: only filter release namespace
As part of upstream Helm chart sanitization, we remove all Namespace
objects, as we create them manually before the release is installed,
otherwise Helm will complain. This is the solution from times, where
charts were not updated to not include or conditionally include the
namespace and Helm (v2) was storing all release objects in kube-system
namespace.

This commit changes this behavior to only filter single namespace
object, which is where we will be installing the release. This patch
aims to fix the installation of charts, which spans across multiple
namespaces like for example Istio.

Right now, many helm charts offers an option to disable namespace
creation already, which we should eventually utilize and disable the
filtering completely, as it is confusing and brings non-obvious results,
especially when comparing 'lokoctl component install' with 'helm
install'. However, this is out of scope for this changeset.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>