pkg/components/cluster-autoscaler: fix checking device uniqueness
Considering following real-life scenario:

- User X creates cluster A named "foo".
- User X creates cluster B also named "foo" (Packet allows that).
- User Y creates cluster C named "bar".

All clusters are created in the same facility.

Now, if user Y tries to install cluster-autoscaler component, it will
fail, as facility have 2 devices with the same name, despite they don't
even belong to user Y clusters.

This is because we check device uniqueness before checking the cluster
name on the devices, which seems incorrect.

This commit breaks down 'getWorkerUserdata()' function into few smaller
functions to make it more readable and to allow implementing simple
tests for the desired behavior.

Also, previously we were returning userData of last node found and now
we return userData of first node, to simplify the code. There is also a
test included for it now.

This commit should also simplify fixing #767.

Closes #766.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>