Refactor Terraform executor
- Define a new type terraform.ExecutionStep which allows describing
  arbitrary Terraform operations declaratively.
- Allow passing one or more ExecutionSteps to the Terraform executor.

This allows making pkg/platform generic as now every platform
implementation can define its steps for deploying infrastructure in a
declarative way, thus making the code which interacts with Terraform
generic. Thanks to the PreExecutionHook concept, the same is true even
when arbitrary logic needs to be executed before some Terraform
operation: A platform implementation can include a callback function
in the PreExecutionHook field of an ExecutionStep and the generic
Terraform code will invoke it at the right time.