pkg/dns: Handle stdin read errors
fmt.Scanln() returns an error which we should handle. fmt.Scanln()
returns an error not only when there was a problem reading from
stdin, but also when a newline is returned on its own, i.e. when the
user hits Enter without typing "skip". Since we rely on the user
hitting Enter for looping here, in order to distinguish between an
actual error and the user hitting Enter, we would have to check for
the sentinel value "unexpected newline" which isn't exported as a
constant by the fmt package.

By using ReadString() we can handle read errors simply without
treating newlines as an error.