refactor: add new method to interface to get provider specific values
This commit moves the helm values.yaml template to a separate file
`template.go` for each provider.

In this case, since the only provider is azure, it is moved to
`azure/template.go`.

This allows for multiple templates, each specific to the provider as
each provider will have provider specific BackupStorageLocation and
VolumeSnapshotLocation.

Values method returns the values specific to the plugin provider. This
is then clubbed together with the common values for the chart such as
metrics for rendering the chart.

Signed-off-by: Imran Pochi <imran@kinvolk.io>