aws-ebs-csi-driver: add NetworkPolicy allowing access to metadata
After we created aws-ebs-csi-driver component, we added a patch to
Lokomotive, which deploys a Global Network Policy, which blocks access
to EC2 Instance Metadata by default for all pods, which ended up
breaking the component functionality.

The issue was not spotted before, as the component does not have
readiness probes defined, which has been reported upstream:
https://github.com/kubernetes-sigs/aws-ebs-csi-driver/issues/548

This commit fixes the component functionality, by adding the
NetworkPolicy object selecting the controller pods, which unblocks all
egress traffic for it, which bypasses the Global Network Policy.

Closes #864

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>