pkg/platform: introduce PlatformWithPostApplyHook interface
To address issue with AKS having delayed creation of default
StorageClass, which affects installing components which depend on the
storage, this commit introduces PlatformWithPostApplyHook interface, which platforms
will be able to implement to obtain kubeconfig file content after cluster
is installed to run their own sanity checks. In case of AKS, hook will
be looping and waiting until the default storage class appears on the
cluster.

Refs #855

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>