pkg/lokomotive: adjust timeouts
3 minutes seems like an arbitrary time to wait for cluster to become
ready, so let's make it at least round to 5 minutes.

As currently we have issues, that nodes do not become ready within 3
minutes timeframe, let's give them max 10 minutes to converge, which
should be sufficient in most cases, as sometimes it may take up to 5
minutes for node to become ready.

Closes #940.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>