packet: Read BGP peer address from metadata service
In some Packet facilities the BGP peer address isn't the same as the
gateway address allocated for a host. Rather, it is a loopback
address that's reachable via the gateway.

The Packet metadata service now exposes BGP info to hosts, so we can
query the metadata service for the BGP peer address.

We currently use the first peer address only since MetalLB doesn't
support multiple node peers yet.

The source address is explicitly specified since when the peer
address is a loopback address, the source IP addresses which ends up
getting selected by the kernel is the node's *public* address which
doesn't work. In cases where the peer address is the gateway address
there is no harm in explicitly specifying the source.

Removing the `/bin/sh -c ""` wrapper in ExecStart because we no
longer need a shell since now we read the peer address from an
environment file.

Removing the TODO about using Afterburn for the peer address since we
no longer use the gateway address as the peer address.

Fixes #1009.