test/ingress/aws: fix buggy test
We weren't taking into account the test structure component name and
were using the hardcoded "httpbin" string. This means we were only
testing the httpbin component.

This uses the component name from the test structure. Also, this changes
the hardcoded "get" subpath to a parameter so we can access the right
Prometheus Operator URL.