gangway: add a serviceaccount
Gangway relies on a service account (any service account) to be mounted
on the pod file system to figure out the cluster CA.

We've disabled automounting of the default service account with the
our admission wehbook and this was breaking Gangway.

This creates a dummy service account so the Gangway pod has access to
the cluster CA file.