velero: remove default configuration in restic
This commit removes the default BackupStorageLocation config in restic
plugin.

Name is optional and its default value is `default` in Helm chart.
Provider is required and making it a default of `aws` doesn't make sense
as the user has to provide  value to it anyway.

Region is also specifc to the provider used. If not `aws` then there is
no need to input that value.

Signed-off-by: Imran Pochi <imran@kinvolk.io>