velero: add provider field to configuration
This commit brings back the addition of `provider` field in the
component configuration. This is done so to be able to provide default
configurations for the plugins.

Example: Without the field, if restic provider is configured, then
`restic.NewConfiguration()` is never set and the provider never gets the
default values.

In future, there may be more fields that are added to the plugins
required custom default values rather than the go default values.

Signed-off-by: Imran Pochi <imran@kinvolk.io>