pkg/platform/packet: set AuthToken field from env var while initializing
Running CCM on Packet requires passing Packet API key to the deployment,
which is currently done trough Terraform, so we can no longer just let
Packet Terraform provider read this environment variable for it's
operation, but we must pass it to Terraform module.

This commit ensures, that if user did not configure auth token, value
from environment variable will be still written into Terraform code and
passed to the controller module.

Signed-off-by: knrt10 <kautilya@kinvolk.io>