prometheus_operator: add new resources for updated prometheus.
NOTE - This commit is to be removed from the branch once a newer version
of Lokomotive post v0.6.0 is released. This commit is added to ensure
backwards compatibility with older Prometheus-operator component by
undoing the changes made in the previous commit, hence it is a temporary
commit in th branch.

This commit creates the following addtional resources:

- Endpoint and Service objects for etcd.
- PrometheusRule object for Contour.
- PrometheusRule object for Rook.
- PrometheusRule object for Metallb.

New resources created above differ in name and the label only offering
compatibility with updated Prometheus Operator. The old resources offer
backward compatibility, meaning if the the cluster is updated etcd
metrics collection won't be broken and if prometheus-operator component
is updated then the metrics collection of contour, rook and metallb wont
be broken even if those components are updated before updating
prometheus-operator.

Old label:
```
app: prometheus-operator
```

New label:
```
app: kube-prometheus-stack
```

Signed-off-by: Imran Pochi <imran@kinvolk.io>