pkg/components/cluster-autoscaler: fix duplicate device error
Packet API seems to have a bug, that when you use Project API key
for querying devices, it returns duplicated entries. However IDs
of the devices should be the same, so it should be safe to relax
the check on those, as if there would actually be 2 different devices
with the same name, they should have different IDs.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>