linkerd: Generate expiry date in `RFC3339` format
This commit changes the expiry date format from the golang's default to
`RFC3339`. This change is introduced because the proxy injector errors
with following errors:

```
failed to inject sidecar. Reason: failed to unmarshal JSON from:
/var/run/linkerd/config/values: error unmarshaling JSON: while decoding
JSON: parsing time "2021-11-09 07:43:12 +0000 UTC" as
"2006-01-02T15:04:05Z07:00": cannot parse " 07:43:12 +0000 UTC" as "T"
```

This problem should be fixed once we update the linkerd libs, and it
would be safe to revert this commit.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>