Add tests for disabling BGP on Packet
This commit adds e2e test, which verifies, that if worker pool has BGP
disabled on Packet, it actually remains disabled.

This test addresses
https://github.com/kinvolk/lokomotive/issues/548#issuecomment-694826634
to ensure, that we do not introduce a regression while adding CCM
support for Packet, as currently CCM enables BGP on all workers in the
project, so it is expected that this test will fail with CCM enabled,
unless CCM is modified to only enable BGP on selected nodes.

Refs #548.

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>