cli/cmd/cluster: load controlplane charts from cluster assets dir
Rather than from lokoctl binary.

This is a prerequisite to be able to restore cluster to the desired
state before starting the upgrade, which is required to safely being
able to add node pools with TLS bootstrap enabled.

The idea is to run 'helm rollback <last release number>' before
extracting new charts to user assets directory or running Terraform, to
ensure that cluster is in the desired and healthy state.

Refs #916

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>