kube-scheduler: fix access to cluster certificates ConfigMap
Currently kube-scheduler is configured to dynamically load cluster certificates from
extension-apiserver-authentication-reader ConfigMap, but it is not authorized to do that
because its ServiceAccount is missing extension-apiserver-authentication-reader Role.

This causes kube-scheduler to not being able to read new cluster certificates dynamically
and makes it print annoying error message.

This commit adds missing RoleBinding for Role extension-apiserver-authentication-reader to
kube-scheduler ServiceAccount to fix this.

Closes: #1096

Signed-off-by: knrt10 <kautilya@kinvolk.io>
Co-authored-by: Mateusz Gozdek <mateusz@kinvolk.io>