deploy-cluster: Don't create ssh key for tinkerbell
This commit adds a check to disable ssh key creation when the platform
is tinkerbell. This is a regressions introduced since moving from ci
based script that created clusters vs the script in lokomotive repo that
creates a cluster.

Without this code the pipeline stays stuck in interactive y/n question
to overwrite the key.

Signed-off-by: Suraj Deshmukh <suraj@kinvolk.io>