aws-ebs-csi-driver: add dynamic provisioning,resizing,snapshot options
This commit enhances the aws-ebs-csi-driver with additional
configuration options such as:
- Volume scheduling for dynamic provisioning
- Volume resizing
- Volme snapshotting

This features are by default enabled and are of great value when
dynamically provisioning volumes on Lokomotive AWS platform.

This also adds functionality for resizing the volumes after they have
been initially provisioned.

Finally enabling volume snapshotting capabilities helps to create volume
snapshots in order to take backup and retore the volumes using Velero
CSI plugin.

Volume snapshot requires additional CRDS, which are now part of the helm
chart in the directory `crds`. The follwing CRDs are created:
 - VolumeSnapshotClasses
 - VolumeSnapshotContents
 - VolumeSnapshots

Signed-off-by: Imran Pochi <imran@kinvolk.io>