baremetal: use grub.cfg file for kernel arguments
While the kernel arguments can be set for PXE, they have not been
applied to the final installation. Also, the Ignition URL got lost
because it was downloaded in the installer and put to
/usr/share/oem/config.ign through the -i ignition.json flag.  Use the
/usr/share/oem/grub.cfg file to permanently set the kernel arguments and
the Ignition URL, so that on each boot the final installation uses the
kernel arguments provided by the user and that Ignition is able to fetch
the any config changes from Matchbox.