velero: add check for restic.region
This commit adds a conditional check for restic.region is to be set if
restic.provider is `aws`.

This change is done to reflect the addtional configuration of region
that is required when provider is `aws`. Without this additional check
lokoctl throws an error message that is caught few levels below in the
stack in the velero. Catching this error allows a consistent error
message reporting to the user.

If any other provider is used, then region value is optional.

Updates the unit test to reflect the same.

Signed-off-by: Imran Pochi <imran@kinvolk.io>