velero: update type of Tolerations.TolerationSeconds
This commit updates the TolerationSeconds from string to int64.
Also update the json tag from `toleration_seconds` to `tolerationSeconds`.

JSON tag is updated because, Tolerations struct is directly marshalled to
JSON and stored as a string TolerationsRaw which is then passed to the
helm templates. This causes the Kubernetes API Server to complain during
validation that no such field regarding `toleration_seconds` since API
server was expected `tolerationSeconds`.

Data type is changed to int64, becuase Kubernetes API server expects the
TolerationSeconds to be int64 instead of string. Thereby causing
validation to fail at the API server end.

Updates the unit test to reflect the same.

Signed-off-by: Imran Pochi <imran@kinvolk.io>