util: add validation to Tolerations
This commit adds validation to the Tolerations type.

This is done so because Kubernetes API server expects that if
`TolerationSeconds` is set then `Effect` must be set to `NoExecute`.
This condition checks for the same and reports and error to the user if
this condition is not met.

Adds unit test for the same.

Signed-off-by: Imran Pochi <imran@kinvolk.io>